Implement drag and drop in MaskPuzzle classes
We no longer use the MouseGrabbable and MouseGrabbableGroup scripts, and no longer do the parent/child swapping hackery.
This was needed to accurately detect which fragment was on top when multiple fragments were stacked under the cursor, and also to use the precise point they were clicked on
(including Z) for dragging.