Set up spin functionality
Set the sprite to spin properly and count its spins. Total number of spins needed to clear the game is set as a public int.

Currently does not technically require spins or spins in a specific direction, but a forced direction will likely be given to the player in the future via RNG as well as a direction that must be spun.