MilkPour: Refactor sprite scaling
Reconfigured the sprites and rewrote the cup code so that win conditions
are 100% accurate to the sprites now, without the need of arbitrary
positional offsets.

The sprites I made are a little bland and could use a bit of an artist's
touch. The actual difficulty may also need some retuning as I may
not have fully translated the numbers from last build.