Microgame/moon soccer/early animations (#393)
* Added the first character animations

Also added lots of new assets to the game

* Added transition between animation states

* Timed the activation of the ball with the kick anim

* Adjusted hitboxes to match new sprites

* Several small changes to the way Kaguya moves

This is a messy commit. She now stops when the ball is kicked and goes into her
anticipation animations. She has a very slight deacceleration when stopping.
Kayuya's start position is now always in front of Mokou. Their init code were
moved to Awake() for that reason.

* Added Kaguya's ball caught sprite

Also adjusted the ball's hitbox to match the new sprite

* Added a small delay before Kaguya stops to catch the ball

* Deleted old unused assets

* Removing more unused files